apple-codesign: add tests for entitlements DER encoding
Code signatures can contain DER encoding of plists representing
entitlements. There is no specification for how Apple is doing the
DER encoding. So we'll need to reverse engineer it by observing
what Apple's tooling does.

This commit adds tests which invoke `codesign` with custom XML
based entitlements. We then capture the DER encoded entitlements
and record the binary representation in our tests. To detect
changes in the future, we compare the content.

This test only runs on macOS since we need a `codesign` executable.
There are currently a handful of compiler warnings on !macOS. But
these will go away once we implement DER encoding in a future commit
and verify that our encoder matches Apple's output.