apple-codesign: implement DER encoding of plists
(This commit was authored by Gregory Szorc but was heavily based on
code written by David Craven submitted to
https://github.com/indygreg/PyOxidizer/pull/515.)

This commit implements low-level functionality for serializing a plist
value to DER.

Tests have been added to verify that our DER encoding matches the values
recorded from identical operations using `codesign`. This hopefully
proves that the implementations are identical.

A future commit will add the DER encoded entitlements to a blob in the
code signature.