apple-codesign: remove incorrect verification error about missing blob for info.plist and resources
The digests are included in the code directory but the data being
digested is external to the signature. So there is no corresponding
blob whose content needs to be digested.

The verifier should ideally verify this external content. But I don't
want to think about implementing that at the moment.

Closes #512.