apple-codesign: implement notarization upload
This commit represents the next step in notarization support: uploading
of content.

We implement support for producing an `.itmsp` directory containing the
zip file and `metadata.xml` that need to be uploaded to Apple. We
support calling into Apple Transporter to automate the upload of these
files. We parse its output and capture the upload ID.

However, this commit stops short of waiting for the notarization
to finish or fetching the result. We have an API to obtain the
notarization ticket already. But I want to implement the missing status
and logs querying pieces before we hook things up end-to-end otherwise
users won't know how to debug failures.