apple-codesign: implement notarization upload status checking
This shores up the code introduced in the previous commit for querying
Apple's servers about the state of an upload.

We can now better interpret the state of an upload as it changes
through opaque processing pipelines. We can also now wait on an
upload to complete by polling the status API.

We are very close to being able to perform a full notarization...