apple-codesign: add an EmbeddedSignatureBuilder for constructing superblobs
This logic is generic and deserves to live outside of the Mach-O
signing code.

As part of this refactor, we make the signature building a bit more
opinionated. For example, slots are now serialized in ascending order
of their slot number.

As part of this, the new type handles calculation digests of blobs/slots
when the code directory is registered. This simplifies some logic
and makes it so the generation of the blobs only occurs once. This
should eliminate some duplicate log output performed by
create_special_blobs().