apple-codesign: implement print-signature-info command
I want to make it easier for people to report buggy behavior with this
tool. The command added in this commit prints out important information
about code signatures and facilitates more easily debugging differences
between `codesign` and this tool.

There are some gaps in coverage. But this is a start.