apple-xar: accommodate multiple stored file names
It looks like productbuild and who knows what else is emitting
multiple <name> elements in XAR table of contents data structures.
This seems flat out wrong. But what can you do?

This commit teaches our XAR ToC parser to handle this. This required
propagating Result through everywhere. Result propagation is probably
not a bad thing. But the reason for it is.