ci: split out common environment setup code
There's a bit of duplicated code in CI YAML. This commit starts the
process of factoring it out.

As part of this, we introduce Just for managing simple shell snippets.

The added secure_download.py script is largely curbed from
python-build-standalone. I have no clue how many times I've written
this code and I really wish curl and wget would offer a mode to validate
the content digest on download.