apple-codesign: implement support for signing with YubiKeys
This commit introduces support for signing with YubiKeys. I've
tested it on my 5c NFC and the cryptography appears to work with
RSA and ECDSA keys/certificates.

I haven't yet tested importing an Apple certificate. But it should
work. I intend to implement automated import of keys into the
smartcard device in a future commit to make it easier to automate
this process without having to go through 3rd party tooling.