apple-bundles: discard nested bundles whose child directories match known dirs
This fixes some issues where immediate child directories of bundles
were improperly flagged as bundles. e.g. a `Resources/Info.plist` would
cause the `Resources/` directory to be flagged as a bundle.

We don't yet handle the case of nested directories. For that, we'll
need to keep track of all bundle directories seen so far and
make the nested scanning aware of the parent.