apple-bundles: more iteration on nested bundles parsing
The recent commit to sniff for relative paths within bundles was
flawed in a few ways. I'm not convinced this code is perfect. But
using recursively calling into `nested_bundles()` so we don't
have to worry about normalizing relative paths from the nearest
parent bundle in the path hierarhy seems like an easier implementation,
albeit at the expense of system calls to perform directory traversal.