apple-codesign: API and display of alternative code directories
As part of debugging bugs with framework signing, I discovered that
the main Mach-O in frameworks appears to always use alternative
code directories.

We don't currently have much code for handling these beyond the slot
definitions. But code directories they are and we know code directories!

This commit adds some APIs for alternative code directory slot/blob
access and it teaches our reading / analysis code to print their
contents. We split out the display normalized version of code
directories to its own type to facilitate reuse.

The results clearly show that frameworks appear to use SHA-1 in the
main code directory but SHA-256 in the first alternative slot. We
have work to do...