apple-codesign: implement own logic for classifying files when traversing bundles
There are bugs in the bundle traversal file identification code
making it difficult to get reliable reports out of our debugging
commands. Rather than rely on the accuracy of this implementation (which
influences bundle signing), let's just implement the heuristic
ourselves. This should hopefully be more resilient to bugs in the
bundle traversal logic.