apple-bundles: better handling of symlinks
Before, we failed to detect a file was a symlink because we tested
against obtained metadata after following the symlink. Doh.

In addition, we weren't emitting symlinks to directories from
files() because we would follow the symlink and think the symlink
was a directory.

And in nested bundles identification, we would follow the directory
symlink to a bundle and emit the symlink path as a bundle.

This commit should make symlink handling consistent.