apple-codesign: better handle versioned frameworks
A framework with a `Versions` directory isn't itself signed. Rather,
each version directory is signed as its own bundle.

This commit teaches the bundle signer to recognize this special case
and only handle the copying of (hopefully just symlink) files when
a versioned framework is encountered.

There are still issues when signing frameworks after this change. But
this commit (and those leading up to it) goes a long way to shore up
framework signing support.