apple-codesign: key off presence of `Resources` directory to determine rules
I'm pretty sure the choice of which set of rules to use boils down
to whether there is a `Resources` directory, not whether the bundle is
shallow or not.

This change does cause our `CodeResources` file to agree with what
`codesign` produces. So I think this is correct, at least in the
case for frameworks.