apple-codesign: use SHA-1 on primary code directories in frameworks
I'm not sure why, but at least framework bundles use SHA-1 on the
primary code directory in the main executable in the bundle.

This commit teaches bundle signing to switch the digest type
to SHA-1 for the main executable of framework bundles.

The diff of signed frameworks between Apple's tooling and our own
is pretty small now...