apple-bundles: stop files traversal for symlinks as well
Before, a symlink was always emitted even if we weren't traversing into
nested directories. After, we obey nested traversal rules always, even
for symlinks.

This fixes a bug in bundle signing where we were attempting to sign
symlinks belonging to a nested bundle. After this change, I'm able to
sign an app bundle with a nested framework which contains symlinks!