apple-codesign: properly look for Resources directory
I broke this a few commits ago when I changed the logic. This needs
to respect whether we're a shallow bundle or not.

After making this change, non-main Mach-O binaries in the complex/nested
bundle I'm using as my test subject are properly signed and their code
directory digests are added to the CodeResources file.

However, our output varies from Apple due to SHA-1 vs SHA-256
again. That should hopefully be easy enough to fix...