apple-codesign: refactor alternative code directory handling
Sleeping on this, I don't think the settings API should be framed in
terms of code directories. Instead, settings should express the
desire to record an alternative digest type and consumers should
translate that to an alternative code directory.

In addition, the low-level details of which slot to occupy should
be abstracted away to avoid externalizing knowledge of the low-level
data structures as much as possible.