apple-codesign: use command line argument to control extra digests
When I was iterating on bundle signing last night, I was trying to get a
particular test bundle to converge with the signatures Apple produced.
The bundle in question was emitting both SHA-1 and SHA-256 digests so I
thought this is what Apple's tooling did by default for all frameworks.

On further examination, frameworks can have just SHA-256 digests. So it
isn't appropriate for us to always sign frameworks with SHA-1 + SHA-256.

Now, there may be other heuristics that warrant opting in to SHA-1
digests automatically. But I have yet to discover those.

This commit refactors teaches the CLI an `--extra-digest` argument to
control allowing extra digests in signatures. Since we can now enable
multi-digest support via the CLI, bundle signing no long mucks with
signing settings to emit multiple digests.

I'm still not thrilled about the overall API here. Perhaps a unified
`Vec<DigestType>` on settings would be more appropriate. But this is
a start.