apple-codesign: implement smart card key generation
Now that we have CSR support implemented and working with Apple
issued certificates, the next step is to support having Apple
issue a certificate associated with a private key that never leaves
the smart card.

This commit implements that.

We added a new `smartcard-generate-key` command for generating a
private key on the smart card. We then taught `smartcard-import` to
allow importing a certificate without touching the private key.

There is some unwanted code to generate a fake certificate to associate
with the wiped key. I'm unsure if Yubikeys don't allow getting access
to the public key info for a slot without a certificate or what.
But it is annoying we can't get the SPKI data except through the
certificate data.