apple-codesign: respect active digests in CodeResources file
Before, we would always emit both the SHA-1 and SHA-256 digest in the
`<files2>` data. I added support for SHA-1 last night because I saw it
was in use.

What I learned today after comparing output is that `CodeResources`
seems to respect the active digest for the overall signing operation.
This makes intuitive sense.

So this commit teaches the `CodeResources` builder to only emit the
digests that are actively in use. Although SHA-1 is always emitted
in `<files>` data. So this only applies to `<files2>` in reality.