apple-codesign: preserve entitlements when signing Mach-O in bundles
We currently only do this for the main executable in the bundle. This
may not be sufficient. But this change does close down a big gap in
the diffs of bundle signing.

Since we have entitlements now, we will emit DER entitlements as well
since that is our default behavior. That may cause differences with
bundles signed with older versions of codesign.