apple-codesign: associate nested bundle digest with parent bundle
Before, we were registering all child bundles' digest with the main/root
bundle. This was wrong. The digest should be associated with the nearest
parent.

This commit fixes that.

We may want to revisit our APIs for nested bundle traversal and expose a
depth first mechanism for emitting bundles. This would likely make the
traversal and recursion logic a bit easier to read. But that's for
another commit.