apple-codesign: start to handle multiple rules set
We are currently merging rules and rules2 into the same rules sets and
only taking the rules2 rules into account when processing a file. This
appears to be wrong as it appears the rules sets are processed
independently.

This commit refactors some code around rules registration to get us
a step closer to processing the rules sets differently.