apple-codesign: traverse into nested bundles when sealing resources
Before, we stopped traversal at nested bundles boundaries. This worked
in most cases. However, the resources builder has its own rules for
determining which files to process and in theory it is capable of making
its own decisions about whether to process a nested file.

I'm worried about the handling of `nested` on rules after this change.
We currently assume that a non-symlink in a `nested` rule is a Mach-O
binary. This appears to be the case in bundles that I'm using the test.
But I'm not convinced it is always true.