apple-codesign: process nested bundles when signing the bundle
Before, we had a pile of code in the main bundle signing loop to
collect the code directory metadata for a child bundle so it could be
stashed away and fed into the signing of its parent bundle. This was
overly complex.

This commit throws all that code away and has the single bundle signing
logic look for nested bundles and pull out code directory digests at
sign time. There is definitely some filesystem I/O overhead involved
here from having to perform multiple scans. But I think the new way
is much easier to understand and closer in spirit to how resources
should be handled.

One thing I don't like is how we're not honoring resource rules when
adding the child bundle. We could address this in a follow-up commit.