apple-codesign: preserve entitlements on all Mach-O binaries in a bundle
Previously, we only preserved the entitlements on the main executable.
Now, we preserve entitlements on all Mach-O binaries.

As part of this we moved the code for importing entitlements into the
signing settings into `SigningSettings` itself. There's probably room
to move other "import state" functionality - possibly to the same
(renamed) function as well since there is a hodgepodge of code that
attempts to carry over settings. I think we should strive for an end
state where settings are canonical and the actual low-level signing
operation doesn't attempt to reconcile state itself.