apple-codesign: preserve binary identifier when signing Mach-O in bundles
Before, we would derive the identifier of not main executable binaries
from their filename. This could be wrong, as there may already be
a valid identifier in the binary to use.

After this change, we import the binary identifier from the existing
Mach-O file.

As part of this, we leverage our existing function for importing Mach-O
signature state into settings and bloat it to cover the binary
identifier.

This fixes yet another difference in behavior between us and codesign.