apple-codesign: record digests of output bundles
This was a stupid bug. When adding nested digests to a bundle, we were
reading the digests from the source bundle/directory and not the output
directory. Essentially, we were using the old digests instead of the new
ones.

With this change, `codesign` is reporting that some complex bundles I'm
testing properly validate without nested bundle verification errors.