apple-codesign: import settings from mach-o data in single mach-o signing mode
We introduced this call to bundle signing several commits ago. I think
it makes sense to keep the logic for reconciling settings state as close
as possible.

This commit teaches the unified signing code path to call this function
when signing a single Mach-O binary.

This might introduce some subtle changes in behavior. This probably
warrants a release note once the refactoring is complete.