apple-codesign: stop carrying forward code resources data
It generally isn't safe to carry forward digests since we may be
targeting a different digest.

This change sounds dangerous and it may break resigning binaries
within bundles. However, binaries arguably shouldn't be signed in
isolation outside the context of a bundle.

If we are signing a bundle, we should provide the code resources data
to the signing settings. So I'm optimistic this won't break things
in practice unless people are looking to shoot themselves in the
foot by signing part of a bundle.

That being said, maybe we do want to recognize when we're signing
the main executable in a bundle and automatically provide the code
resources data? If so, we should handle that in the calling code,
before we get to low-level signing code.