apple-codesign: import runtime version into settings
As part of making settings canonical. There is still logic for
handling the runtime version in low-level signing. Ideally this
would also be moved into import_settings_from_macho().

And with this commit, we no longer consult the old code directory when
performing low-level signing operations: the settings are mostly
canonical!