apple-codesign: implement remote code signing
I've been wanting to implement this feature for months.

My initially conceived use case for this is enabling a low trust CI
worker to delegate signing operations to a human using a hardware
device for signing. But there are other similar use cases, such as
a team of developers and only a few people have access to the
signing key.

This commit includes an implementation of the websocket server
including a Terraform module to deploy it to AWS. I have this
running in a personal account and it seems to be pretty stable.

There is tons of documentation for the feature, including extensive
documentation of the protocol and cryptosystem design.

Since I rolled my own crypto here, my guess is there are some
weaknesses with it. Issue #552 tracks having it audited by... someone.
Until then, the docs have a giant warning about the potential for
security issues. Even if there are critical weaknesses in the
cryptosystem, signing requires presence of a signer. It's not like
the private key is sitting around on a CI worker waiting to be
exfiltrated by anyone who can run a PR or anyone with machine
level access to those workers. So I'd like to think that even if
there is a weakness here that other design choices limit the
potential for danger. Of course, I still very much want the
security to be sound!

There's a ton of small paper cuts remaining with the implementation.
Lots of small things we can do to improve security, like prompting
for consent before signing operations, keeping a "transparency log"
of signing operations, etc. But at this point I've probably spent
~30 hours on this feature and just want to get something merged.

Partially implements #387.