apple-codesign: render more CMS signed attributes
When debugging why signatures weren't validating with Apple, I stumbled
upon a difference in the handling of the CMS signed attributes data.
I thought this problem space was solved long ago. But apparently there
are still bugs here.

Capturing the CMS signed attributes data in our signature rendering
clearly clues us into the problems.

As part of this, we also print the SHA-1 of blobs because CMS attributes
use SHA-1.