apple-codesign: include all CD digests in CMS attributes
Before, we only included a single digest in the CMS attributes.
After, we include all code directory digests.

With this change, Apple's tools are now validating signatures on
binaries with multiple code directories. This scenario is much more
common than before since we now automatically add a 2nd code directory
if a binary is targeting an older OS version.