apple-codesign: properly record symlinks in notarization zip files
This works around lack of symlink support in the zip crate. I added
this feature upstream yesterday. But it will take a while to release
and we're stuck on old versions of the zip crate for dependency hell
reasons. So it might be a while before we can upgrade.

After this change, I'm able to successfully notarize bundles containing
symlinks. That's ~every bundle containing a framework, since framework
versions are often symlinks.