apple-codesign: support encoding SJS as PEM
When attempting to remotely sign on GitHub Actions, the CI worker
output was truncating long lines. We could reduce the length of the
SJS. Or we could express it differently to get around length
limitations.

This commit does the latter and introduces PEM encoding of session
join strings. This requires opening an editor to allow saving the
multiline value.