pyoxy: build more variations in CI
We now build Python 3.8 variants of pyoxy. No 3.10 for the moment
because of a known crasher (which should hopefully be fixed by the
end of a weekend after a PyO3 release).

We also lay the groundwork for x86_64-unknown-linux-musl builds. But
they aren't yet produced due to a build issue.