pyembed: expand documentation of interpreter types
I'm working on a mechanism to export Rust source docs to rst so we can
have a single source of docs for these fields and can reference them in
the Sphinx docs. In preparation for this, bring the documentation of
the Rust types up to our standards, as we'll be deleting Sphinx
documentation once the new feature is ready.