python-packaging: refactor component licensing handling
We refactor LicensedComponent so ComponentFlavor has the name embedded.
This is more appropriate, as it doesn't require us to shoehorn a name
into every component. It also enables us to index components by
type, without potential for name conflicts.

As part of this, we overhaul the license report and how it is displayed.

We are still far from my desired end state with license handling. But
this is a big step in the right direction.