python-packaging-net: delete crate
This crate was recently renamed. It existed mainly to allow us to
fetch SPDX license texts. Turns out that the spdx crate has an optional
feature to expose the texts. So that means we no longer need to perform
HTTP requests to resolve license text. That means we can roll
functionality in this crate into python-packaging. We started doing
this in the previous commit.