release: use guppy for deriving licenses
We are currently emitting dev dependencies in the licenses list.
`cargo deny` doesn't give us the control we need to filter out
unwanted dependencies. So this commit switches us to guppy so we
can traverse the dependency graph is custom semantics.

We try to preserve existing behavior as part of the conversion.