tugger-rust-toolchain: avoid fetching remote manifest is possible
Before we were incurring network requests even if we already had
the toolchain locally. This could lead to failures because networks
aren't reliable.

After this commit, we only incur the network request if we need to
resolve the manifest content. For cases where we already have the
toolchain installed, this makes toolchain resolution faster and
more reliable.