pyoxidizer: refactor handling of licensing file
It is better for this logic to live in the embedded context type because
it is more generic.

A side-effect of this is that the licensing file is initially
registered during embedded context generation. Its content can be
overwritten during Rust project building. But if all we're doing
is deriving the embedded context, the file should be generated
with licensing metadata that is current at the time the embedded
context is derived.