pyoxidizer: canonicalize cargo manifest path and set cwd
Windows CI is erroring due to path normalization issues. I can't
manage to reproduce this locally. I think the issue has something to
do with CWD and the manifest path being on different drives.

Since we're invoking `cargo metadata` under the hood, we should be
executing it in the context of the project under evaluation. But in
order to get the directory reliably, we need to make it absolute.
So we canonicalize the manifest path.