pyoxidizer: add empty [workspace] to generated Rust projects
Today, if I attempt to build an auto-generated Rust project inside
the pyoxidizer checkout or any other workspace, Cargo complains because
the project isn't tracked by the workspace.

In most cases, the Rust projects are ephemeral and aren't attached
to a larger workspace. So I think it makes sense for them to advertise
themselves as isolated by default.