Use new REST API for notarization
This replaces the notarization codepath to use the new REST API made
available by Apple in WWDC22. This API is entirely documented, and meant
for direct user consumption[0].

As a result, this commit also removes the codepath using the Transporter
API for notarization, and thus the requirement for the iTMSTransporter
tool to be present on the machine for notarization.

Note that the API has some undocumented bits that were discovered
through trial and error. One such bit is that the S3 token that apple
returns to upload files are valid for the us-west-2 region.

[0]: https://developer.apple.com/documentation/notaryapi/submitting_software_for_notarization_over_the_web