cryptographic-message-syntax: fix sorting of signed attributes
The `SignedAttributes` set must be sorted according to DER rules.
Our previous implementation of sorting sorted by the OID of each
attribute. This was only correct in some scenarios.

After this commit, we sort by the DER encoding of the full attribute.
This ensures that duplicate attributes and attributes with different
encoding lengths are sorted properly.

Part of #614.