apple-codesign: copy data between Mach-O segments
Before, Mach-O segments that had padding between segments crashed
due to an assertion. Having padding between Mach-O segments is
allowed - albeit not observed until recently.

So this commit teaches the segment writing code to preserve the space
between segments when writing out a Mach-O.

We still fail to handle these gaps when computing code signatures.
This will be addressed in a separate commit.

Part of #588 and #616.