linux-package-analyzer: disable RPM reading
It saddens me that it has come to this. But the rpm-rs crate doesn't
appear to have an active maintainer. Its old dependencies are making
it difficult to keep our dependencies fresh. So we have little choice
but to cut our ties so we can modernize.

Issue #619 tracks restoring the functionality. And we keep the code
around to remind us of the sadness.