global: RustCrypto updates
yubikey.rs 0.6 was recently released. This release enables us to take
upgrades for various RustCrypto crates, finally enabling us to get on
the latest versions of many of these crates.

This requires various API changes. Most notably, the owned parsed
document types for various ASN.1 data structures went away. So we
now store a validated `SecretDocument` and introduce functions for
converting to a lightly parsed type with borrowed data to facilitate
usage. There's probably better ways to do this. But this is the
easiest path forward.