global: move apple-bundles, apple-codesign, apple-flat-packages, and apple-xar out of repository
The canonical home of these projects is now
https://github.com/indygreg/apple-platform-rs.

This commit deletes the crates from this repo and updates / removes
various references to these projects as appropriate.

As part of this, we transition crate dependencies to published versions
of these crates. We just published new versions of these crates from
their new home so project metadata in the latest published version is
accurate and to pick up new crate versions for other crates recently
moved out of this repository. This was necessary to avoid diamond
dependency problems.