global: move tugger-file-manifest to own repository
In order to extract the apple-* crates to their own repository, we
need to also extract some other crates in order to avoid diamond
dependency problems.

The most annoying crate with the diamond dependency problem is
tugger-file-manifest, which is used in tons of crates.

I recently extract this crate to its own repository at
https://github.com/indygreg/simple-file-manifest-rs and published
version 0.11.0.

This commit updates all references to `tugger-file-manifest` to use the
renamed and externally published `simple-file-manifest` crate.