global: move cryptography-message-syntax, pgp-cleartext, x509-certificate to own repository
I'm splitting up the monorepo.

In order to move the `apple-*` crates out of this repo, we need to move
crates they depend on as well in order to avoid the diamond dependency
problem.

I have moved the `cryptographic-message-syntax`, `pgp-cleartext`, and
`x509-certificate` crates to
https://github.com/indygreg/cryptography-rs.

This commit deletes those crates from this repository and updates
references to use the latest published versions, which should be
identical to the in-repo versions previously referenced.